% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_table.R
\name{adjacency_table}
\alias{adjacency_table}
\title{Find links between edges based on number of neighbors}
\usage{
adjacency_table(x, level = 1, self = FALSE)
}
\arguments{
\item{x}{a \code{character} vector of edges labels.}

\item{level}{an \code{integer} of length 1. The number of neighbors used to
define vertices.}

\item{self}{a \code{logical} of length 1. If \code{TRUE}, a site can be linked to
itself. Default is \code{FALSE}.}
}
\value{
A \code{data.frame} with three columns:
\itemize{
\item \code{from}: label of one of the two edges of the vertice (link)
\item \code{to}: label of the other edge of the vertice
\item \code{weight}: the presence or absence of vertice between two edges
}
}
\description{
Finds links between edges based on number of neighbors. The edges (argument
\code{x}) will be ordered to find links in a directional way (from upstream to
downstream).

\strong{IMPORTANT:} The order of sites must be found in the sites labels.
}
\examples{
\dontrun{
## Import Adour sites ----

path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "bridge")
                            
adour_sites <- read.csv(path_to_file)


## Select the 5 first sites ----

adour_sites <- adour_sites[1:5, ]


## Create adjacency table with 1 degree of neighborhood ----

adjacency_df <- adjacency_table(adour_sites$"site")
}
}
