% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_edges_lists.R
\name{append_edges_lists}
\alias{append_edges_lists}
\title{Append several edges lists}
\usage{
append_edges_lists(...)
}
\arguments{
\item{...}{one or several edges lists \code{data.frame}. Outputs of the function
\code{\link[=create_edges_list]{create_edges_list()}}.}
}
\description{
Appends several edges lists created by \code{\link[=create_edges_list]{create_edges_list()}}. Merged edges
will be ordered and duplicates will be removed.
}
\examples{
library("chessboard")

# Two-dimensional sampling (only) ----
sites_infos <- expand.grid("transect" = 1:3, "quadrat" = 1:5)

nodes <- create_node_labels(data     = sites_infos, 
                            transect = "transect", 
                            quadrat  = "quadrat")

edges_1 <- create_edges_list(nodes, method = "pawn", directed = TRUE)
edges_2 <- create_edges_list(nodes, method = "bishop", directed = TRUE)

edges <- append_edges_lists(edges_1, edges_2)
}
