% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge-package.R
\docType{package}
\name{bridge-package}
\alias{bridge}
\alias{bridge-package}
\title{bridge: Directional Spatial Network Analyses}
\description{
Contains different tools used by the FRB-CESAB Project Bridge to work with directional spatial network.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/bridge}
  \item Report bugs at \url{https://github.com/frbcesab/bridge/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
