% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_to_int}
\alias{char_to_int}
\title{Convert a character to an integer}
\usage{
char_to_int(x)
}
\arguments{
\item{x}{a character vector. Must contain numbers.}
}
\value{
An integer vector of the same length as \code{x}.
}
\description{
Converts a character to an integer by removing non-number characters
(letters, punctuation, special characters, ...).

\strong{Important:} Use this function with caution: \code{"AD11"} will result in \code{11}
and \code{"A1D1"}, \code{"1A-D1"}, etc. will also result in \code{11}.
}
\note{
This function is no more in use.
}
