% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chessboard-package.R
\docType{package}
\name{chessboard-package}
\alias{chessboard}
\alias{chessboard-package}
\title{chessboard: Neighborhood and Connectivity in Spatial Networks}
\description{
Contains different tools used by the FRB-CESAB Project Bridge to work with directional spatial network.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/chessboard}
  \item Report bugs at \url{https://github.com/frbcesab/chessboard/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
