% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nodes_labels.R
\name{create_nodes_labels}
\alias{create_nodes_labels}
\title{Create unique nodes labels}
\usage{
create_nodes_labels(transects, quadrats)
}
\arguments{
\item{transects}{a \code{numeric} vector. The identifier of the transects. If
missing, a unique transect will be created and named \code{1}
(for the purpose of the package).}

\item{quadrats}{a \code{numeric} vector. The identifier of the quadrats. If
missing, a unique quadrat will be created and named \code{1}.
(for the purpose of the package).}
}
\value{
A \code{data.frame} with the three following columns:
\itemize{
\item \code{node}, the nodes label (created by the function)
\item \code{transect}, the transects label (created by the function or provided)
\item \code{quadrat}, the quadrats label (created by the function or provided)
}
}
\description{
Creates unique nodes (sampling units) labels in (un)directed two-dimensional
(spatial) networks (i.e. regular grid).

It's important to note that, even the package \code{bridge} is designed to deal
with spatial networks, it does not deal explicitly with spatial coordinates.
\strong{But nodes labels matters!}

\strong{Convention}

The package \code{bridge} requires that the sampling has two dimensions: one
from bottom to top (called \strong{quadrats}), and one from left to right (called
\strong{transects}). If the sampling has been conducted along one single
dimension (\strong{transects} or \strong{quadrats}), this function will create a
fictitious label for the missing dimension.

In other words, the package \code{bridge} can work with sampling designs such as
regular grids (two dimensions), transects (one dimension), and quadrats
(one dimension).

\strong{Nodes labeling}

The nodes labels will be of the form: \code{1-2}, where \code{1} is the identifier of
the transect (created by the function is missing), and \code{2}, the identifier
of the quadrat (created by the function is missing).
}
\examples{
library("bridge")

# Two-dimensional sampling ----
sites_infos <- expand.grid("transect" = 1:3, "quadrat" = 1:5)
sites_infos

nodes <- create_nodes_labels(transects = sites_infos$"transect", 
                             quadrats  = sites_infos$"quadrat")
nodes

gg_chessboard(nodes)

# One-dimensional sampling (only transects) ----
sites_infos <- 1:5

nodes <- create_nodes_labels(transects = sites_infos)
nodes

gg_chessboard(nodes)

# One-dimensional sampling (only quadrats) ----
sites_infos <- 1:5

nodes <- create_nodes_labels(quadrats = sites_infos)
nodes

gg_chessboard(nodes)
}
