% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_unique_key}
\alias{create_unique_key}
\title{Create an unique node identifier}
\usage{
create_unique_key(data)
}
\arguments{
\item{data}{a \code{data.frame}. Each column will be used to create an unique
edge identifier.}
}
\value{
A character vector with nodes labels.
}
\description{
Create an unique node identifier
}
\note{
This function is no more in use.
}
