% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_matrix.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{Convert a data frame to an adjacency matrix}
\usage{
df_to_matrix(data, lower = TRUE, upper = TRUE, diag = TRUE)
}
\arguments{
\item{data}{a \code{data.frame} with the following three columns: \code{from} (edge),
\code{to} (edge), and \code{weight} (value of the vertice).}

\item{lower}{a logical value. If \code{TRUE} (default), keep values in the lower
triangle of the matrix. Otherwise their will be replaced by \code{NA}.}

\item{upper}{a logical value. If \code{TRUE} (default), keep values in the upper
triangle of the matrix. Otherwise their will be replaced by \code{NA}.}

\item{diag}{a logical value. If \code{TRUE} (default), keep values in the
diagonal of the matrix. Otherwise their will be replaced by \code{NA}.}
}
\value{
An integer vector of the same length as \code{x}.
}
\description{
Converts a data frame to an adjacency matrix.
}
\examples{
\dontrun{
## _Add an example_
}
}
