% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges_weight_matrix.R
\name{edges_weight_matrix}
\alias{edges_weight_matrix}
\title{Create an edges weight matrix}
\usage{
edges_weight_matrix(distances, lower = TRUE, upper = TRUE, diag = TRUE)
}
\arguments{
\item{distances}{a \code{data.frame} with the following three columns: \code{from}
(the first node of the edge), \code{to} (the second node of the edge), and
\code{weight} (the weight of the edge between the two nodes, e.g. a distance).}

\item{lower}{a logical value. If \code{TRUE} (default), keep values in the lower
triangle of the matrix. Otherwise they will be replaced by \code{NA}.}

\item{upper}{a logical value. If \code{TRUE} (default), keep values in the upper
triangle of the matrix. Otherwise they will be replaced by \code{NA}.}

\item{diag}{a logical value. If \code{TRUE} (default), keep values in the
diagonal of the matrix. Otherwise they will be replaced by \code{NA}.}
}
\value{
An edges weight matrix of dimensions \verb{n x n}, where \code{n} is the
number of nodes (sites).
}
\description{
Creates an edges weight matrix from the output of \code{\link[=distance_euclidean]{distance_euclidean()}}.
}
\examples{
\dontrun{
# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "chessboard")
adour_sites <- read.csv(path_to_file)

# Convert sites to sf object (POINTS) ----
adour_sites <- sf::st_as_sf(adour_sites, coords = 2:3, crs = "epsg:2154")

# Compute distances between pairs of sites along the Adour river ----
adour_dists <- distance_euclidean(adour_sites)

# Create Edges weight matrix ----
edges_weight_matrix(adour_dists)

# Create Edges weight matrix (with options) ----
edges_weight_matrix(adour_dists, lower = FALSE)
edges_weight_matrix(adour_dists, upper = FALSE)
edges_weight_matrix(adour_dists, diag = FALSE)
}
}
