% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbors.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Find neighbors among a list of nodes}
\usage{
find_neighbors(
  nodes,
  focus,
  method,
  degree = 1,
  directed = FALSE,
  reverse = FALSE,
  self = FALSE
)
}
\arguments{
\item{nodes}{a \code{data.frame} with (at least) the following three columns:
\code{node}, \code{transect}, and \code{quadrats}. Must be the output if the function
\code{\link[=create_nodes_labels]{create_nodes_labels()}}.}

\item{focus}{an \code{character} of length 1. The node label for which the
neighbors must be found. Must exist in the \code{nodes} object.}

\item{method}{a \code{character} of length 1.}

\item{degree}{an \code{integer} of length 1. The maximum number of neighbors to
search for in one direction.}

\item{directed}{a \code{logical} of length 1. If \code{FALSE} (default), search for
neighbors in all directions (undirected network). Otherwise, the network
will be considered as directed according to the main direction of the
network (i.e. through quadrats).}

\item{reverse}{a \code{logical} of length 1. If \code{TRUE}, change the orientation of
the network (i.e. through quadrats). This argument is ignored if
\code{directed = FALSE}.}

\item{self}{a \code{logical} of length 1. If \code{TRUE}, a node can be its own
neighbor. Default is \code{FALSE}.}
}
\value{
...
}
\description{
Find neighbors among a list of nodes
}
