% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_chessboard.R
\name{gg_chessboard}
\alias{gg_chessboard}
\title{Plot a sampling as a chessboard}
\usage{
gg_chessboard(nodes, xlab = "Transect", ylab = "Quadrat")
}
\arguments{
\item{nodes}{a \code{data.frame} with the following three columns:
\code{node}, \code{transect}, and \code{quadrats}. See \code{\link[=create_nodes_labels]{create_nodes_labels()}} for
further information.}

\item{xlab}{a \code{character} of length 1. The title of the top axis.
Default is \code{'Transect'}.}

\item{ylab}{a \code{character} of length 1. The title of the left axis.
Default is \code{'Quadrat'}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots a sampling as a chessboard of dimensions \code{t} x \code{q}, with
\code{t}, the number of transects, and
\code{q}, the number of quadrats.
}
\examples{
# Two-dimensional sampling ----
sites_infos <- expand.grid("transect" = 1:3, "quadrat" = 1:5)

nodes <- create_nodes_labels(transects = sites_infos$"transect", 
                             quadrats  = sites_infos$"quadrat")

gg_chessboard(nodes)

# One-dimensional sampling (only transects) ----
sites_infos <- 1:5

nodes <- create_nodes_labels(transects = sites_infos)

gg_chessboard(nodes)

# One-dimensional sampling (only quadrats) ----
sites_infos <- 1:5

nodes <- create_nodes_labels(quadrats = sites_infos)

gg_chessboard(nodes)
}
