% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_to_points.R
\name{line_to_points}
\alias{line_to_points}
\title{Convert a LINESTRING to POINT}
\usage{
line_to_points(x, density = 0.01, type = "regular", ...)
}
\arguments{
\item{x}{an \code{sf} object of type \code{LINESTRING}.}

\item{density}{a \code{numeric} of length 1. The density of points to sample on
the linear structure. See \code{\link[sf:st_line_sample]{sf::st_line_sample()}} for further detail.
Default value is \code{1/100}.}

\item{type}{a character of length 1. Either \code{regular} (default) or \code{random}.
Default is \code{regular}.}

\item{...}{other argument to pass to  \code{\link[sf:st_line_sample]{sf::st_line_sample()}}.}
}
\description{
Converts a spatial line of type \code{LINESTRING} into \code{POINT} by sampling points
on a linear geometry using the function \code{\link[sf:st_line_sample]{sf::st_line_sample()}}. This function
is used by the function \code{\link[=distance_along]{distance_along()}}.
}
\examples{
\dontrun{
## Import Adour river (LINESTRING) ----

path_to_file <- system.file("extdata", "adour_lambert93.gpkg", 
                            package = "bridge")
                            
adour_river <- sf::st_read(path_to_file)

## Sample points of the Adour river ----

adour_river_pts <- line_to_points(adour_river)
}
}
