% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_links.R
\name{make_links}
\alias{make_links}
\title{Find links between edges based on number of neighbors}
\usage{
make_links(x, level = 1)
}
\arguments{
\item{x}{a character vector of edges labels.}

\item{level}{an integer. The number of neighbors used define vertices.}
}
\value{
A \code{data.frame} with three columns:
\itemize{
\item \code{from}: label of one of the two edges of the vertice (link)
\item \code{to}: label of the other edge of the vertice
\item \code{weight}: the presence or absence of vertice between two edges
}
}
\description{
Finds links between edges based on number of neighbors. The edges (argument
\code{x}) will be ordered to find links in a directional way (from upstream to
downstream). This order must be found in the edges labels.
}
\examples{
\dontrun{
## _Add an example_
}
}
