% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_df.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{Convert an adjacency matrix to a data frame}
\usage{
matrix_to_df(mat, na_rm = TRUE)
}
\arguments{
\item{mat}{a \code{matrix} object. The adjacency table to be converted.}

\item{na_rm}{a logical value. If \code{TRUE} (default), removes \code{NA} vertices.}
}
\value{
A \code{data.frame} with three columns:
\itemize{
\item \code{from}: label of one of the two edges of the vertice (link)
\item \code{to}: label of the other edge of the vertice
\item \code{weight}: the presence or absence of vertice between two edges
}
}
\description{
Converts an adjacency matrix to a data frame
}
\examples{
\dontrun{
## _Add an example_
}
}
