% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_by_edges_matrix.R
\name{nodes_by_edges_matrix}
\alias{nodes_by_edges_matrix}
\title{Create a nodes-by-edges matrix}
\usage{
nodes_by_edges_matrix(edges)
}
\arguments{
\item{edges}{a \code{data.frame} with the following three columns: \code{from} (the
first node of the edge), \code{to} (the second node of the edge), and \code{edge}
(presence/absence of the edge).}
}
\value{
A list of two elements:
\itemize{
\item \code{se.mat}: the nodes-by-edges matrix of dimensions \verb{n x k}, where \code{n} is
the number of nodes and \code{k} the number of edges (including the edge
between the fictitious origin and the first site);
\item \code{edges}: a \code{data.frame} of edges list.
}
}
\description{
Creates a nodes-by-edges matrix that will be used by \code{\link[adespatial:aem]{adespatial::aem()}}.
This function creates the same output as \code{\link[adespatial:aem.build.binary]{adespatial::aem.build.binary()}}
but works in a different way: it's only based on nodes labels (not on
coordinates). Also, this function adds labels to nodes and edges.
}
\examples{
# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "bridge")
adour_sites  <- read.csv(path_to_file)

# Retrieve nodes (from nodes vector) ----
adour_nodes <- nodes_list(adour_sites$"site")

# Find edges with 1 degree of neighborhood ----
adour_edges <- edges_list(adour_nodes, directed = TRUE)

# Create nodes-by-edges matrix ----
adour_matrix <- nodes_by_edges_matrix(adour_edges)
adour_matrix
}
