% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_list.R
\name{nodes_list}
\alias{nodes_list}
\title{Retrieve and order nodes from an edges list}
\usage{
nodes_list(edges)
}
\arguments{
\item{edges}{a \code{data.frame} with at least two columns:
\itemize{
\item \code{from}: label of one of the two nodes of the edge
\item \code{to}: label of the other node of the edge
}}
}
\value{
A \code{data.frame} with column, \code{node} containing unique ordered nodes
labels.
}
\description{
Creates labels for edges (non edges are ignored). If two edges are identical
(e.g. \code{"S01-S02"} and \code{"S02-S01"}), there will have the same label
(\code{"E-01"}).
}
\examples{
# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "bridge")
adour_sites  <- read.csv(path_to_file)

# Find edges with 1 degree of neighborhood (undirected network) ----
edges <- edges_list(adour_sites$"site")

# Retrieve nodes ----
nodes_list(edges)
}
