% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot a connectivity or a nodes-by-edges matrix}
\usage{
plot_matrix(x, title)
}
\arguments{
\item{x}{a \code{matrix} object. An adjacency or nodes-by-edges matrix.}

\item{title}{a \code{character} of length 1. The caption of the figure.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots an adjacency or a nodes-by-edges matrix.
}
\examples{
# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "bridge")
adour_sites  <- read.csv(path_to_file)

# Retrieve nodes (from nodes vector) ----
adour_nodes <- nodes_list(adour_sites$"site")

# Find edges with 1 degree of neighborhood (undirected network) ----
adour_edges <- edges_list(adour_nodes, directed = TRUE)

# Get adjacency matrix ----
adour_adj_matrix <- adjacency_matrix(adour_edges)

# Visualize matrix ----
plot_matrix(adour_adj_matrix, title = "Connectivity matrix")

# Nodes by edges matrix ----
adour_edges_mat <- nodes_by_edges_matrix(adour_edges)

# Visualize matrix ----
plot_matrix(adour_edges_mat$"se.mat", title = "Nodes-by-edges matrix")
}
