% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_units.R
\name{recode_units}
\alias{recode_units}
\title{Create an unique edge identifier}
\usage{
recode_units(data)
}
\arguments{
\item{data}{a \code{data.frame}. Each column will be used to create an unique
edge identifier.}
}
\value{
A character vector with edges labels.
}
\description{
\emph{Add an good description}
}
\examples{
\dontrun{
## _Add an example_
}
}
