% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chessboard-package.R
\docType{package}
\name{chessboard-package}
\alias{chessboard}
\alias{chessboard-package}
\title{chessboard: Neighborhood and Connectivity in Spatial Networks}
\description{
Provides functions to work with directed (asymmetric) and undirected (symmetrical) spatial networks. It makes easier the creation of connectivity matrix, i.e. a binary matrix of dimensions n x n, where n is the number of nodes (sampling units) indicating the presence (1) or the absence (0) of an edge (link) between pairs of nodes. Different network objects can be produced by chessboard: nodes list, neighbors list, edges list, connectivity matrix.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/chessboard}
  \item Report bugs at \url{https://github.com/frbcesab/chessboard/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
