% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chessboard-package.R
\docType{package}
\name{chessboard-package}
\alias{chessboard}
\alias{chessboard-package}
\title{chessboard: Create Network Connections Based on Chess Moves}
\description{
Provides functions to work with directed (asymmetric) and undirected (symmetric) spatial networks. It makes the creation of connectivity matrix easier, i.e. a binary matrix of dimension n x n, where n is the number of nodes (sampling units) indicating the presence (1) or the absence (0) of an edge (link) between pairs of nodes. Different network objects can be produced by 'chessboard': node list, neighbor list, edge list, connectivity matrix. It can also produce objects that will be used later in Moran's Eigenvector Maps (Dray et al. (2006) \doi{10.1016/j.ecolmodel.2006.02.015}) and Asymetric Eigenvector Maps (Blanchet et al. (2008) \doi{10.1016/j.ecolmodel.2008.04.001}), methods available in the package 'adespatial' (Dray et al. (2023) \url{https://CRAN.R-project.org/package=adespatial}). This work is part of the FRB-CESAB working group Bridge \url{https://www.fondationbiodiversite.fr/en/the-frb-in-action/programs-and-projects/le-cesab/bridge/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/chessboard}
  \item Report bugs at \url{https://github.com/frbcesab/chessboard/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Erica Rievrs Borges \email{erica.rievrs@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0001-7751-6265}{ORCID})
  \item Eric Tabacchi \email{eric.tabacchi@univ-tlse3.fr} (\href{https://orcid.org/0000-0001-7729-4439}{ORCID})
  \item Guillaume Fried \email{guillaume.fried@anses.fr} (\href{https://orcid.org/0000-0002-3653-195X}{ORCID})
  \item Nicolas Mouquet \email{nicolas.mouquet@cnrs.fr} (\href{https://orcid.org/0000-0003-1840-6984}{ORCID})
}

}
\keyword{internal}
