% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_along.R
\name{distance_along}
\alias{distance_along}
\title{Compute the distance between two nodes along a linear shape}
\usage{
distance_along(
  sites,
  along,
  density = 0.01,
  type = "regular",
  mc.cores = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{sites}{an \code{sf} object of type \code{POINT}. A spatial object
containing coordinates of sites. Note that the first column must be the
site label.}

\item{along}{an \code{sf} object of type \code{LINESTRING}. A spatial object
containing coordinates of the linear shape (e.g. a river, a road, etc.)
to follow while computing distances between two nodes (sites).}

\item{density}{a \code{numeric} of length 1. The density of points to sample on
the linear structure. See \code{\link[sf:st_line_sample]{sf::st_line_sample()}} for further detail.
Default is \code{1/100}.}

\item{type}{a \code{character} of length 1. The method to sample points on the
linear shape. Either \code{regular} (default) or \code{random}.}

\item{mc.cores}{an \code{integer} of length 1. The number of cores to use
(must be lesser than the number of cores available on the machine obtained
with \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{...}{other argument to pass to  \code{\link[sf:st_line_sample]{sf::st_line_sample()}}.}
}
\value{
A three-column \code{data.frame} with:
\itemize{
\item \code{from}, the first node
\item \code{to}, the second node
\item \code{weight}, the distance between the two nodes along the linear shape
}
}
\description{
Computes distance between two nodes (sites) along a linear shape
(e.g. a river, a road, etc.). This function uses parallel computing to speed
up the distance computation.
}
\examples{
# Import Adour river (LINESTRING) ----
path_to_file <- system.file("extdata", "adour_lambert93.gpkg", 
                            package = "chessboard")
adour_river <- sf::st_read(path_to_file)

# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "chessboard")
adour_sites <- read.csv(path_to_file)

# Convert sites to sf object (POINTS) ----
adour_sites <- sf::st_as_sf(adour_sites, coords = 2:3, crs = "epsg:2154")

\dontrun{
# Compute distances between pairs of sites along the Adour river ----
dist_along_river <- distance_along(adour_sites, adour_river)
}
}
