% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges_list.R
\name{edges_list}
\alias{edges_list}
\title{Find edges between nodes based on a degree of neighborhood}
\usage{
edges_list(x, level = 1, self = FALSE)
}
\arguments{
\item{x}{a \code{character} vector of nodes (sites) labels.}

\item{level}{an \code{integer} of length 1. The number of neighbors used to
define vertices.}

\item{self}{a \code{logical} of length 1. If \code{TRUE}, a site can be linked to
itself. Default is \code{FALSE}.}
}
\value{
A \code{data.frame} with three columns:
\itemize{
\item \code{from}: label of one of the two edges of the vertice (link)
\item \code{to}: label of the other edge of the vertice
\item \code{weight}: the presence or absence of vertice between two edges
}
}
\description{
Finds edges (links) between nodes (sites) based on a degree of neighborhood.
The nodes (argument \code{x}) will be ordered to find edges in a directional way
(from upstream to downstream along a linear shape).

\strong{IMPORTANT:} The order of sites must be found in the sites labels.
}
\examples{
\dontrun{
# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "bridge")
adour_sites  <- read.csv(path_to_file)

# List of edges with 1 degree of neighborhood ----
edges_list(adour_sites$"site")
}
}
