% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges_to_sf.R
\name{edges_to_sf}
\alias{edges_to_sf}
\title{Convert edges to LINESTRING spatial object}
\usage{
edges_to_sf(edges, sites)
}
\arguments{
\item{edges}{a \code{data.frame} with four columns:
\itemize{
\item \code{edge_id}: label of the edge
\item \code{edge}: 0 (no edge) or 1 (edge)
\item \code{from}: label of one of the two nodes of the edge
\item \code{to}: label of the other node of the edge
}}

\item{sites}{an \code{sf} object of type \code{POINT}. A spatial object with
coordinates of sites. Note that the first column must be the sites labels.}
}
\value{
An \code{sf} spatial object of type \code{LINESTRING} where the number of rows
correspond to the number of edges.
}
\description{
Convert edges to an \code{sf} spatial object of type \code{LINESTRING} with one row
per edge.
}
\examples{
# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "bridge")
adour_sites  <- read.csv(path_to_file)

# Find edges with 1 degree of neighborhood ----
adour_edges <- edges_list(adour_sites$"site")

# Convert sites to spatial POINT ----
adour_sites_sf <- sf::st_as_sf(adour_sites, coords = 2:3, crs = "epsg:2154")

# Convert edges to spatial LINESTRING ----
edges_sf <- edges_to_sf(adour_edges, adour_sites_sf)
head(edges_sf)

# Visualization ----
plot(sf::st_geometry(adour_sites_sf), pch = 19)
plot(sf::st_geometry(edges_sf), add = TRUE)

# Find edges with 2 degrees of neighborhood ----
adour_edges <- edges_list(adour_sites$"site", degree = 2)

# Convert sites to spatial POINT ----
adour_sites_sf <- sf::st_as_sf(adour_sites, coords = 2:3, crs = "epsg:2154")

# Convert edges to spatial LINESTRING ----
edges_sf <- edges_to_sf(adour_edges, adour_sites_sf)
head(edges_sf)

# Visualization ----
plot(sf::st_geometry(adour_sites_sf), pch = 19)
plot(sf::st_geometry(edges_sf), add = TRUE)
}
