% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_edges_list.R
\name{matrix_to_edges_list}
\alias{matrix_to_edges_list}
\title{Convert an adjacency matrix to an edges list}
\usage{
matrix_to_edges_list(x, all = FALSE)
}
\arguments{
\item{x}{a \code{matrix} object. The adjacency matrix to be converted in edges
list.}

\item{all}{a logical value. If \code{FALSE} (default), removes missing edges.}
}
\value{
A \code{data.frame} with four columns:
\itemize{
\item \code{edge_id}: label of the edge
\item \code{edge}: 0 (no edge) or 1 (edge)
\item \code{from}: label of one of the two nodes of the edge
\item \code{to}: label of the other node of the edge
}
}
\description{
Converts an adjacency matrix to an edges list. This function allows to create
the same edges list as the one obtained with \code{\link[=edges_list]{edges_list()}}.
}
\examples{
# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "bridge")
adour_sites  <- read.csv(path_to_file)

# Retrieve nodes (from nodes vector) ----
adour_nodes <- nodes_list(adour_sites$"site")

# Find edges with 1 degree of neighborhood (undirected network) ----
adour_edges <- edges_list(adour_nodes)
adour_edges

# Get adjacency matrix ----
adour_adj_matrix <- adjacency_matrix(adour_edges)
adour_adj_matrix

# Convert back to edges list ----
new_edges <- matrix_to_edges_list(adour_adj_matrix)
new_edges

# Check ----
identical(adour_edges, new_edges)
}
